# examples/t1api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModuleListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "modules": {
            "items": {
              "$ref": "#/components/schemas/ModuleItem"
            },
            "title": "Modules",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "modules",
          "count",
          "request_id"
        ],
        "title": "ModuleListResponse",
        "type": "object"
      },
...
      "RouteRequest": {
        "properties": {
          "automatic_fallback": {
            "default": false,
            "description": "Only meaningful with model_id set: fall through the cascade if the requested model is exhausted, instead of raising MODEL_QUOTA_EXHAUSTED.",
            "title": "Automatic Fallback",
            "type": "boolean"
          },
          "input_tokens": {
            "default": 0,
            "minimum": 0.0,
            "title": "Input Tokens",
            "type": "integer"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Manual model selection. Omit for automatic routing.",
            "title": "Model Id"
          },
          "plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional assertion of the caller's plan. The server resolves the real plan from the key's assignment; a mismatch is rejected, never silently accepted.",
            "title": "Plan"
          }
        },
        "title": "RouteRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Controlled gateway into HyperNix-pip. The client requests an operation; the server decides what exists, what's available, and how much is left. See wiki/T1-API.md for the full contract.",
    "title": "HyperNix T1 API",
    "version": "0.71.5b3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/models/route": {
      "post": {
        "description": "Not in the original endpoint list \u2014 added to give the routing/\nquota-cascade engine (``t1api.routing``) an HTTP surface, since the\nspec describes the engine's behavior in detail but doesn't enumerate\nan endpoint for it. See wiki/T1-API.md#routing.\n\nAutomatic routing (no ``model_id``) walks the caller's plan cascade.\nManual routing (``model_id`` set) enforces the exhausted-model rule\nexactly as the spec describes: ``MODEL_QUOTA_EXHAUSTED`` unless\n``automatic_fallback=True``.\n\n**Beta 3 change, deliberately breaking Beta 2's shape:** the plan is\nno longer read from the request body. It is resolved from the key's\nserver-side assignment (:meth:`t1api.keys.KeyDirectory.resolve_plan`),\nand a ``plan`` in the body is treated as an assertion \u2014 matching is\nfine, mismatching is refused. Beta 2 let a client name its own plan,\nwhich is the exact thing the design principle forbids: \"the server\ndetermines... which models the user can access, which limits apply,\nwhich fallback models are allowed.\"\n\nA key narrowed to a model subset (``allowed_models`` on its\nassignment) is also checked here, so manual selection can't reach a\nmodel the assignment excludes even when the registry and quota would\nboth allow it.",
        "operationId": "route_model_models_route_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route Model",
        "tags": [
          "models"
        ]
      }
    },
...
    "/modules": {
      "get": {
        "operationId": "list_modules_modules_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Modules",
        "tags": [
          "modules"
        ]
      }
    },
    "/modules/create": {
      "post": {
        "operationId": "create_module_modules_create_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Module",
        "tags": [
          "modules"
        ]
      }
    },
    "/modules/receive": {
      "post": {
        "description": "Beta 3: accept a signed module push from another T1 server.\n\nAuthenticated by the shared deployment secret rather than by a T1\nkey \u2014 a peer server is not a user, holds no key of its own, and\ngiving every deployment target an admin key just to receive bytes\nwould be a much larger grant than this needs.\n\nOrder matters and is enforced here: verify the signature (which\ncovers the body digest, method, path, and a freshness window), then\nverify the declared checksum against the bytes actually received,\nand only then store. Nothing is written to disk before both checks\npass, and nothing is ever executed.",
        "operationId": "receive_module_modules_receive_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-T1-Signature",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-T1-Signature",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-T1-Timestamp",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-T1-Timestamp",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-T1-Content-SHA256",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-T1-Content-Sha256",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-T1-Module-ID",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-T1-Module-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleReceiveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Receive Module",
        "tags": [
          "modules"
        ]
      }
    },
    "/modules/upload/local": {
      "post": {
        "operationId": "upload_local_modules_upload_local_post",
        "parameters": [
          {
            "in": "query",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_local_modules_upload_local_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Local",
        "tags": [
          "modules"
        ]
      }
    },
    "/modules/upload/remote": {
      "post": {
        "description": "Validates and *registers* a remote source \u2014 does not fetch it (see\nt1api/modules.py's module docstring). The module is marked\nPENDING_FETCH; actually staging the content is a separate concern not\nwired to a job kind in this beta (no real remote fetch infrastructure\nexists to safely exercise yet \u2014 see wiki/T1-API.md#roadmap).",
        "operationId": "upload_remote_modules_upload_remote_post",
        "parameters": [
          {
            "in": "query",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleUploadRemoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Remote",
        "tags": [
          "modules"
        ]
      }
    },
    "/modules/{module_id}": {
      "delete": {
        "description": "Delete a module and its stored bytes.\n\nDestructive and irreversible, so it needs ``?confirm=true`` (see\n``deps.require_confirmation`` / ``T1_REQUIRE_DESTRUCTIVE_CONFIRMATION``)\nand is audited whether it succeeds or is refused.",
        "operationId": "delete_module_modules__module_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Module",
        "tags": [
          "modules"
        ]
      },
      "get": {
        "operationId": "get_module_modules__module_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Module",
        "tags": [
          "modules"
        ]
      },
      "patch": {
        "operationId": "update_module_modules__module_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Module",
        "tags": [
          "modules"
        ]
      }
    },
...
    "/modules/{module_id}/fetch": {
      "post": {
        "description": "Beta 3: stage a module's registered remote source into local storage.\n\nThe URL is not a parameter \u2014 it comes from the registry entry that\n``POST /modules/upload/remote`` validated. Accepting a URL here would\nreopen the SSRF hole that registering it through the validated path\ncloses.",
        "operationId": "fetch_module_source_modules__module_id__fetch_post",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleFetchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch Module Source",
        "tags": [
          "modules"
        ]
      }
    },
...
  }
}